(BSR)[PRO] refactor: remove useAdapter hook
