+++ pro/src/pages/CollectiveOfferVisibility/CollectiveOfferEditionVisibility.tsx
@@ -1,15 +1,17 @@
 import React from 'react'
 import { useNavigate } from 'react-router-dom'
-import { useSWRConfig } from 'swr'
+import useSWR, { useSWRConfig } from 'swr'
 
 import { GetCollectiveOfferResponseModel } from 'apiClient/v1'
 import { AppLayout } from 'app/AppLayout'
 import { CollectiveOfferLayout } from 'components/CollectiveOfferLayout/CollectiveOfferLayout'
-import { GET_COLLECTIVE_OFFER_QUERY_KEY } from 'config/swrQueryKeys'
+import {
+  GET_COLLECTIVE_OFFER_QUERY_KEY,
+  GET_EDUCATIONAL_INSTITUTIONS_QUERY_KEY,
+} from 'config/swrQueryKeys'
 import { isCollectiveOfferTemplate, Mode } from 'core/OfferEducational/types'
 import { computeURLCollectiveOfferId } from 'core/OfferEducational/utils/computeURLCollectiveOfferId'
 import { extractInitialVisibilityValues } from 'core/OfferEducational/utils/extractInitialVisibilityValues'
-import { useAdapter } from 'hooks/useAdapter'
 import useNotification from 'hooks/useNotification'
 import { CollectiveOfferVisibilityScreen } from 'screens/CollectiveOfferVisibility/CollectiveOfferVisibility'
 import {
@@ -18,8 +20,8 @@ import {
 } from 'screens/OfferEducational/useCollectiveOfferFromParams'
 import Spinner from 'ui-kit/Spinner/Spinner'
 
-import { getEducationalInstitutionsAdapter } from './adapters/getEducationalInstitutionsAdapter'
 import { patchEducationalInstitutionAdapter } from './adapters/patchEducationalInstitutionAdapter'
+import { getEducationalInstitutions } from './getEducationalInstitutions'
 
 const CollectiveOfferVisibility = ({
   offer,
@@ -28,11 +30,12 @@ const CollectiveOfferVisibility = ({
   const notify = useNotification()
   const navigate = useNavigate()
   const { mutate } = useSWRConfig()
-  const {
-    error,
-    data: institutionsPayload,
-    isLoading,
-  } = useAdapter(getEducationalInstitutionsAdapter)
+
+  const educationalInstitutionsQuery = useSWR(
+    [GET_EDUCATIONAL_INSTITUTIONS_QUERY_KEY],
+    () => getEducationalInstitutions(),
+    { fallbackData: [] }
+  )
 
   if (isCollectiveOfferTemplate(offer)) {
     throw new Error(
@@ -57,12 +60,8 @@ const CollectiveOfferVisibility = ({
     notify.success(message)
   }
 
-  if (isLoading) {
+  if (educationalInstitutionsQuery.isLoading) {
     return <Spinner />
-  }
-
-  if (error) {
-    return null
   }
 
   return (
@@ -76,8 +75,8 @@ const CollectiveOfferVisibility = ({
             offer.teacher
           )}
           onSuccess={onSuccess}
-          institutions={institutionsPayload.institutions}
-          isLoadingInstitutions={isLoading}
+          institutions={educationalInstitutionsQuery.data}
+          isLoadingInstitutions={educationalInstitutionsQuery.isLoading}
           offer={offer}
         />
       </CollectiveOfferLayout>
